// src/gen/api/types/GetCommentLikesPreview.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { LikesPreview } from './LikesPreview.ts'

export type GetCommentLikesPreviewPathParams = {
  /**
   * @description UUID identifier
   * @type string, uuid
   */
  comment_id: string
}

/**
 * @description Top 5 most recent likers + total
 */
export type GetCommentLikesPreview200 = {
  /**
   * @type object
   */
  data: LikesPreview
}

/**
 * @description Validation failed
 */
export type GetCommentLikesPreview400 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Missing or invalid access token
 */
export type GetCommentLikesPreview401 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Not found
 */
export type GetCommentLikesPreview404 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Internal server error
 */
export type GetCommentLikesPreview500 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

export type GetCommentLikesPreviewQueryResponse = GetCommentLikesPreview200

export type GetCommentLikesPreviewQuery = {
  Response: GetCommentLikesPreview200
  PathParams: GetCommentLikesPreviewPathParams
  Errors:
    | GetCommentLikesPreview400
    | GetCommentLikesPreview401
    | GetCommentLikesPreview404
    | GetCommentLikesPreview500
}