// src/gen/api/types/ReactionUserSummary.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ReactionUserSummary = {
  /**
   * @description UUID identifier
   * @type string, uuid
   */
  id: string
  /**
   * @type string
   */
  firstName: string
  /**
   * @type string
   */
  lastName: string
  /**
   * @type string,null, uri
   */
  avatarUrl: string | null
}